+++ changelog.md
@@ -12,6 +12,14 @@ These changes will (most likely) be included in the next version.
 
 ## [Unreleased]
 
+## [0.104.1] - 2019-12-31
+- It is no longer necessary to have recurrent waves for an arena to work. MobArena automatically creates a "catch all" recurrent wave in case the arena session reaches a wave number that isn't covered by any other wave definitions.
+- Entities outside of the arena can no longer target players, pets, or monsters inside of the arena.
+- Tab completion for `/ma kick` and `/ma restore` now uses actual player names instead of display names.
+- If the world of an exit warp, leaderboard, or linked class chest is not available on (re)load, MobArena now throws a config error instead of failing silently later down the road.
+- Overlapping arena regions are now reported as warnings in the server log during arena load, because overlapping regions can result in undefined, buggy behavior.
+- MobArena's internal version checker has been rewritten. It now uses the resource API of Spigot instead of DBO. It's also a lot more lightweight and caches results for up to one hour.
+
 ## [0.104] - 2019-08-08
 - Extended and upgraded potions are now supported in the item syntax by prepending `long_` or `strong_` to the data portion of a potion item (e.g. `potion:strong_instant_heal:1` will yield a Potion of Healing II). Check the wiki for details.
 - MobArena now has basic tab completion support for most of the commands that take arguments.
@@ -139,7 +147,8 @@ Thanks to:
 - Swatacular for help with testing bug fixes
 - Haileykins for contributions to the code base
 
-[Unreleased]: https://github.com/garbagemule/MobArena/compare/0.104...HEAD
+[Unreleased]: https://github.com/garbagemule/MobArena/compare/0.104.1...HEAD
+[0.104.1]: https://github.com/garbagemule/MobArena/compare/0.104...0.104.1
 [0.104]: https://github.com/garbagemule/MobArena/compare/0.103.2...0.104
 [0.103.2]: https://github.com/garbagemule/MobArena/compare/0.103.1...0.103.2
 [0.103.1]: https://github.com/garbagemule/MobArena/compare/0.103...0.103.1